Add LexicalAnalyzer, rename Remus.{Parsing => TypeParsing}
