--- Remus/Parsing/StringParser.cs
@@ -1,5 +1,0 @@
-﻿namespace Remus.Parsing {
-    public class StringParser {
-        
-    }
-}
