+++ Remus/TypeParsing/StringParser.cs
@@ -1,0 +1,5 @@
+﻿namespace Remus.TypeParsing {
+    internal sealed class StringParser : ITypeParser<string> {
+        public string Parse(string input) => input;
+    }
+}
